Refactor: Extract to ManyJoinTableMethods
